item repository almost implemented
