+++ app/src/main/java/se/miun/dt170/antonsskafferi/ui/bong/BongItemView.java
@@ -1,6 +1,7 @@
 package se.miun.dt170.antonsskafferi.ui.bong;
 
 import android.content.Context;
+import android.util.AttributeSet;
 import android.view.LayoutInflater;
 
 import androidx.annotation.NonNull;
@@ -10,8 +11,8 @@ import se.miun.dt170.antonsskafferi.R;
 
 public class BongItemView extends ConstraintLayout {
 
-    public BongItemView(@NonNull Context context) {
-        super(context);
+    public BongItemView(@NonNull Context context, AttributeSet attrs) {
+        super(context, attrs);
 
         LayoutInflater inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
         inflater.inflate(R.layout.bong_item, this, true);
