+++ app/src/main/java/se/miun/dt170/antonsskafferi/ui/bong/BongListFragment.java
@@ -1,9 +1,11 @@
 package se.miun.dt170.antonsskafferi.ui.bong;
 
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ListView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -19,10 +21,10 @@ import se.miun.dt170.antonsskafferi.R;
  * Displayed to the right in the order {@link se.miun.dt170.antonsskafferi.ui.order_overview.OrderOverviewFragment}.
  * Displayed in the {@link se.miun.dt170.antonsskafferi.activity.KitchenActivity}.
  */
-
 public class BongListFragment extends Fragment {
 
     private BongListViewModel mViewModel;
+    private ListView bongListView;
 
     public static BongListFragment newInstance() {
         return new BongListFragment();
@@ -31,18 +33,23 @@ public class BongListFragment extends Fragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
-
         return  inflater.inflate(R.layout.bong_list_fragment, container, false);
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        bongListView = view.findViewById(R.id.bongListView);
     }
 
     @Override
     public void onActivityCreated(@Nullable Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
         mViewModel = ViewModelProviders.of(this).get(BongListViewModel.class);
-        // TODO: Use the ViewModel
-
-
-
+        mViewModel.bongItems.observe(getViewLifecycleOwner(), items -> {
+            // TODO: Update UI on item changes
+            Log.d("BongListFragment", items.toString());
+            bongListView.addFooterView(new BongItemView(items));
+        });
     }
-
 }
